added logic to depends on view to change based on state
