+++ src/renderer/styles/d3Wrapper.scss
@@ -17,6 +17,6 @@
 
   .depends-wrapper {
     height: 500px;
-    width: 500px;
+    width: 600px;
   }
 }
